+++ spec/open_meteo/forecast/variables_spec.rb
@@ -1,6 +1,7 @@
 RSpec.describe OpenMeteo::Forecast::Variables do
-  let(:variables) { described_class.new(current:, hourly:, daily:, models:) }
+  let(:variables) { described_class.new(current:, minutely_15:, hourly:, daily:, models:) }
   let(:current) { [] }
+  let(:minutely_15) { [] }
   let(:hourly) { [] }
   let(:daily) { [] }
   let(:models) { [] }
@@ -8,8 +9,15 @@ RSpec.describe OpenMeteo::Forecast::Variables do
   describe "#to_get_params" do
     subject { variables.to_get_params }
 
+    context "when no variables are set" do
+      subject { described_class.new.to_get_params }
+
+      it { is_expected.to eq({}) }
+    end
+
     context "when all variables are empty" do
       let(:current) { [] }
+      let(:minutely_15) { [] }
       let(:hourly) { [] }
       let(:daily) { [] }
       let(:models) { [] }
@@ -19,6 +27,7 @@ RSpec.describe OpenMeteo::Forecast::Variables do
 
     context "when current is set" do
       let(:current) { %i[something other] }
+      let(:minutely_15) { [] }
       let(:hourly) { [] }
       let(:daily) { [] }
       let(:models) { [] }
@@ -26,8 +35,19 @@ RSpec.describe OpenMeteo::Forecast::Variables do
       it { is_expected.to eq({ current: "something,other" }) }
     end
 
+    context "when minutely_15 is set" do
+      let(:current) { [] }
+      let(:minutely_15) { %i[something other] }
+      let(:hourly) { [] }
+      let(:daily) { [] }
+      let(:models) { [] }
+
+      it { is_expected.to eq({ minutely_15: "something,other" }) }
+    end
+
     context "when hourly is set" do
       let(:current) { [] }
+      let(:minutely_15) { [] }
       let(:hourly) { %i[something other] }
       let(:daily) { [] }
       let(:models) { [] }
@@ -37,6 +57,7 @@ RSpec.describe OpenMeteo::Forecast::Variables do
 
     context "when daily is set" do
       let(:current) { [] }
+      let(:minutely_15) { [] }
       let(:hourly) { [] }
       let(:daily) { %i[something other] }
       let(:models) { [] }
@@ -46,6 +67,7 @@ RSpec.describe OpenMeteo::Forecast::Variables do
 
     context "when models is set" do
       let(:current) { [] }
+      let(:minutely_15) { [] }
       let(:hourly) { [] }
       let(:daily) { [] }
       let(:models) { %i[something other] }
@@ -55,12 +77,14 @@ RSpec.describe OpenMeteo::Forecast::Variables do
 
     context "when all are set" do
       let(:current) { %i[something other] }
+      let(:minutely_15) { %i[something other] }
       let(:hourly) { %i[something other] }
       let(:daily) { %i[something other] }
       let(:models) { %i[something other] }
       let(:expected_output) do
         {
           current: "something,other",
+          minutely_15: "something,other",
           hourly: "something,other",
           daily: "something,other",
           models: "something,other",