+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Add retry logic to the client
 - Add models to forecast variables
+- Add minutely_15 to forecast variables
 
 ## [0.2.0] - 2023-11-08